Add NSManagedObjectContextExtensions
