--- Stepic/Sources/Extensions/CoreData/NSManagedObjectContext+BatchDeleteRequest.swift
@@ -1,10 +1,0 @@
-import CoreData
-
-extension NSManagedObjectContext {
-    func executeAndMergeChanges(using batchDeleteRequest: NSBatchDeleteRequest) throws {
-        batchDeleteRequest.resultType = .resultTypeObjectIDs
-        let result = try execute(batchDeleteRequest) as? NSBatchDeleteResult
-        let changes: [AnyHashable: Any] = [NSDeletedObjectsKey: result?.result as? [NSManagedObjectID] ?? []]
-        NSManagedObjectContext.mergeChanges(fromRemoteContextSave: changes, into: [self])
-    }
-}
